refactor(auth): move auth config into src/auth and cleanup test scaffold
